Update LoginController.cs
CreateNewUser previously would use the first part of the email address for LoginId. The LoginController would use the last part of User.Identity.Name. This only works if the users samAccountName is the same as the users email address- which isn't always true. This also updates the LoginController to use the full User.Identity.Name