Update UserController.cs
Chaning CreateNewUser to use the samAccountName instead of the email address for the LoginId of a new user. This matches what the LoginController uses to look up a user. Previously, this would only work if a users samAccountName matched their email address- this is not allowed in highly regulated enviroments, as it discloses users samAccountName.